change vlen from u8 to u16 in kernel IR

+++ zyx/src/backend/c.rs
@@ -160,7 +160,7 @@ impl CDevice {
 
         // --- Phase 2: RC and dtype analysis (same as OpenCL) ---
         let mut rcs: Map<OpId, u32> = Map::with_capacity_and_hasher(kernel.ops.len().into(), BuildHasherDefault::new());
-        let mut dtypes: Map<OpId, (DType, u8)> = Map::with_capacity_and_hasher(100, BuildHasherDefault::new());
+        let mut dtypes: Map<OpId, (DType, u16)> = Map::with_capacity_and_hasher(100, BuildHasherDefault::new());
 
         let mut op_id = kernel.head;
         while !op_id.is_null() {
@@ -176,7 +176,7 @@ impl CDevice {
                     dtypes.insert(op_id, (dtype, 1));
                 }
                 &Op::Load { src, index, vlen: len } => {
-                    dtypes.insert(op_id, (dtypes[&src].0, len as u8));
+                    dtypes.insert(op_id, (dtypes[&src].0, len as u16));
                     *rcs.entry(index).or_insert(0) += 1;
                 }
                 &Op::Store { dst, x: src, index, vlen: _ } => {
@@ -205,7 +205,7 @@ impl CDevice {
                 }
                 Op::Vectorize { ops } => {
                     let dtype = dtypes[&ops[0]];
-                    dtypes.insert(op_id, (dtype.0, ops.len() as u8));
+                    dtypes.insert(op_id, (dtype.0, ops.len() as u16));
                     for &x in ops {
                         *rcs.entry(x).or_insert(0) += 1;
                     }
@@ -232,7 +232,7 @@ impl CDevice {
 
         // --- Phase 3: Codegen ---
         let mut reg_map: Map<OpId, usize> = Map::with_capacity_and_hasher(kernel.ops.len().into(), BuildHasherDefault::new());
-        let mut registers: Vec<((DType, u8), u32, u8)> = Vec::new();
+        let mut registers: Vec<((DType, u16), u32, u8)> = Vec::new();
         let mut constants: Map<OpId, Constant> = Map::with_capacity_and_hasher(100, BuildHasherDefault::new());
         let mut indices: Map<OpId, u8> = Map::with_capacity_and_hasher(20, BuildHasherDefault::new());
 
@@ -306,8 +306,8 @@ impl CDevice {
         fn new_reg(
             op_id: OpId,
             reg_map: &mut Map<OpId, usize>,
-            registers: &mut Vec<((DType, u8), u32, u8)>,
-            dtype: (DType, u8),
+            registers: &mut Vec<((DType, u16), u32, u8)>,
+            dtype: (DType, u16),
             rc: u32,
             current_loop_level: u8,
         ) -> usize {
@@ -330,7 +330,7 @@ impl CDevice {
             constants: &Map<OpId, Constant>,
             indices: &Map<OpId, u8>,
             reg_map: &Map<OpId, usize>,
-            registers: &mut [((DType, u8), u32, u8)],
+            registers: &mut [((DType, u16), u32, u8)],
             loop_level: u8,
         ) -> String {
             if let Some(c) = constants.get(&op_id) {
